Removed hard-coded apis directory file so new apis can be picked up without modifying anything.
