+++ lib/generator.js
@@ -31,8 +31,6 @@ var DISCOVERY_URL = argv['discovery-url'] ? argv['discovery-url'] : (
   args.length ? args[0] : 'https://www.googleapis.com/discovery/v1/apis/'
 );
 
-var API_INDEX = './apis/index.js';
-var API_INDEX_TEMPLATE = './templates/api-index.js';
 var API_TEMPLATE = './templates/api-endpoint.js';
 var BEAUTIFY_OPTIONS = {
   'indent_size': 2,
@@ -53,7 +51,6 @@ var BEAUTIFY_OPTIONS = {
 };
 var RESERVED_PARAMS = ['resource', 'media', 'auth'];
 var templateContents = fs.readFileSync(API_TEMPLATE, { encoding: 'utf8' });
-var indexTemplContents = fs.readFileSync(API_INDEX_TEMPLATE, { encoding: 'utf8' });
 var transporter = new DefaultTransporter();
 
 /**
@@ -242,39 +239,6 @@ Generator.prototype.generateAPI = function(apiDiscoveryUrl, callback) {
       uri: apiDiscoveryUrl
     }, _generate);
   }
-};
-
-/**
- * Generate the API index file
- * @param  {Function} callback Callback called with err, filename
- * @throws {Error} If there is an error generating the index file.
- */
-Generator.prototype.generateIndex = function(callback) {
-  var headers = this.options.includePrivate ? {} : { 'X-User-Ip': '0.0.0.0' };
-  transporter.request({
-    uri: DISCOVERY_URL,
-    headers: headers
-  }, function(err, resp) {
-    if (err) {
-      return handleError(err, callback);
-    }
-    var contents = beautify(swig.render(indexTemplContents, { locals: resp }), BEAUTIFY_OPTIONS);
-    mkdirp(path.dirname(API_INDEX), function(err) {
-      if (err) {
-        return handleError(err, callback);
-      }
-      fs.writeFile(API_INDEX, contents, function(err) {
-        if (err) {
-          return handleError(err, callback);
-        }
-        this.log('API index generated in ' + API_INDEX);
-        if (callback && typeof callback === 'function') {
-          callback(null, API_INDEX);
-          return;
-        }
-      }.bind(this));
-    }.bind(this));
-  }.bind(this));
 };
 
 /**
