+++ lib/googleapis.js
@@ -21,12 +21,60 @@
  * @module google
  */
 
+var path = require('path');
+var fs = require('fs');
+var util = require('util');
+
 /**
  * Load the apis from apis index file
  * This file holds all version information
  * @private
  */
-var apis = require('../apis');
+var apis = {};
+
+/**
+ * Return a Function that requires an API from the disk
+ * @param  {String} filename Filename of API
+ * @return {function}        function used to require the API from disk
+ * @private
+ */
+function requireAPI(filename) {
+  return function(options) {
+    var type = typeof options;
+    var version;
+    if (type === 'string') {
+      version = options;
+      options = {};
+    } else if (type === 'object') {
+      version = options.version;
+      delete options.version;
+    } else {
+      throw new Error('Argument error: Accepts only string or object');
+    }
+    try {
+      var endpointPath = path.join(__dirname, filename, path.basename(version));
+      var Endpoint = require(endpointPath);
+      var ep = new Endpoint(options);
+      ep.google = this; // for drive.google.transporter
+      return Object.freeze(ep); // create new & freeze
+    } catch (e) {
+      throw new Error(util.format('Unable to load endpoint %s("%s"): %s',
+        filename, version, e.message));
+    }
+  };
+}
+
+/**
+ * @callback callback
+ * @param {Error} err Error object if an error occurred.
+ * @param {object} data Response data object.
+ */
+
+// Dynamically discover available APIs
+fs.readdirSync(path.join(__dirname, '../apis')).forEach(function(file) {
+  apis[file] = requireAPI('../apis/' + file);
+});
+
 var GoogleAuth = require('google-auth-library');
 
 /**
